requirements.txt
aiocryptopay==0.4.8aiofiles==24.1.0aiogram==3.22.0aiohappyeyeballs==2.6.1aiohttp==3.12.15aiosignal==1.4.0annotated-doc==0.0.3annotated-types==0.7.0anyio==4.11.0APScheduler==3.11.0asyncpg==0.30.0attrs==25.4.0certifi==2023.11.17click==8.3.0colorama==0.4.6dnspython==2.8.0email-validator==2.3.0et_xmlfile==2.0.0fastapi==0.120.0fastapi-cli==0.0.14fastapi-cloud-cli==0.3.1ffmpeg-python==0.2.0frozenlist==1.8.0future==1.0.0greenlet==3.2.4h11==0.16.0httpcore==1.0.9httptools==0.7.1httpx==0.28.1idna==3.10itsdangerous==2.2.0Jinja2==3.1.6loguru==0.7.3magic-filter==1.0.12markdown-it-py==4.0.0MarkupSafe==3.0.3mdurl==0.1.2multidict==6.7.0numpy==2.3.4openpyxl==3.1.5orjson==3.11.4pandas==2.3.3pillow==12.0.0propcache==0.4.1pyaes==1.6.1pyasn1==0.6.1pydantic==2.11.10pydantic-extra-types==2.10.6pydantic-settings==2.11.0pydantic_core==2.33.2Pygments==2.19.2python-dateutil==2.9.0.post0python-dotenv==1.1.1python-multipart==0.0.20pytz==2025.2PyYAML==6.0.3rich==14.2.0rich-toolkit==0.15.1rignore==0.7.1rsa==4.9.1sentry-sdk==2.42.1shellingham==1.5.4six==1.17.0sniffio==1.3.1SQLAlchemy==2.0.43starlette==0.48.0StrEnum==0.4.15Telethon==1.41.2typer==0.20.0typing-inspection==0.4.2typing_extensions==4.15.0tzdata==2025.2tzlocal==5.3.1ujson==5.11.0urllib3==2.5.0uvicorn==0.38.0watchfiles==1.1.1websockets==15.0.1win32_setctime==1.2.0yarl==1.22.0requests~=2.32.5